Crate re_sorbet

Rerun arrow metadata and record batch definitions.

Handles the structure of arrow record batches and their meta data for different use cases for Rerun.

An arrow record batch that follows a specific schema is called a SorbetBatch.

Some SorbetBatches has even more constrained requirements, such as ChunkBatch and DataframeBatch.

Every ChunkBatch is a SorbetBatch.

Every DataframeBatch is a SorbetBatch.

NOTE: DataframeBatch has not yet been implemented.

Each batch type has a matching schema type:

SorbetBatch has a SorbetSchema

ChunkBatch has a ChunkSchema

DataframeBatch will have a DataframeSchema

Modules

chunk_batch 🔒

chunk_schema 🔒

column_descriptor 🔒

column_descriptor_ref 🔒

column_kind 🔒

component_column_descriptor 🔒

error 🔒

index_column_descriptor 🔒

ipc 🔒

metadata 🔒

row_id_column_descriptor 🔒

sorbet_batch 🔒

sorbet_columns 🔒

sorbet_schema 🔒

Structs

ChunkBatch

The [ArrowRecordBatch] representation of a Rerun chunk.

ChunkSchema

The parsed schema of a Rerun chunk, i.e. multiple columns of data for a single entity.

ComponentColumnDescriptor

Describes a data/component column, such as Position3D, in a dataframe.

IndexColumnDescriptor

Describes a time column, such as log_time.

MissingFieldMetadata

MissingMetadataKey

RowIdColumnDescriptor

Describes the schema of the primary [RowId] column.

SorbetBatch

Any rerun-compatible [ArrowRecordBatch].

SorbetColumnDescriptors

SorbetSchema

The parsed schema of a SorbetBatch.

UnsupportedTimeType

WrongDatatypeError

Enums

BatchType

The type of SorbetBatch.

ColumnDescriptor

ColumnDescriptorRef

ColumnError

ColumnKind

The type of column in a sorbet batch.

MismatchedChunkSchemaError

SorbetError

Traits

MetadataExt

Make it more ergonomic to work with arrow metadata.

Functions

ipc_from_schema

Encode an arrow schema as IPC bytes.

schema_from_ipc

Decode an arrow schema from IPC bytes.

Type Aliases

ArrowBatchMetadata

Arrow metadata for an arrow record batch.

ArrowFieldMetadata

Arrow metadata for a column/field.