Crate re_sorbet

Rerun arrow metadata and record batch definitions.

Handles the structure of arrow record batches and their meta data for different use cases for Rerun.

Modules

column_schema 🔒

data_column_schema 🔒

index_column_schema 🔒

ipc 🔒

metadata 🔒

Structs

ComponentColumnDescriptor

Describes a data/component column, such as Position3D, in a dataframe.

MissingFieldMetadata

MissingMetadataKey

TimeColumnDescriptor

Describes a time column, such as log_time.

UnsupportedTimeType

Enums

ColumnDescriptor

ColumnError

Traits

MetadataExt

Make it more ergonomic to work with arrow metadata.

Functions

ipc_from_schema

Encode an arrow schema as IPC bytes.

schema_from_ipc

Decode an arrow schema from IPC bytes.

Type Aliases

ArrowBatchMetadata

Arrow metadata for an arrow record batch.

ArrowFieldMetadata

Arrow metadata for a column/field.